# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Manager API specification
 * OpenAPI spec version: 0.1.0
 */
export interface ErrorResponse {
  error: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface Owner {
  id: string;
  gymId: string;
  name: string;
  gymName: string;
  phone: string;
  email: string;
}

export interface RegisterOwnerRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  gymName: string;
  phone?: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginOwnerRequest {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthResponse {
  token: string;
  owner: Owner;
}

export interface UpdateProfileRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name?: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  gymName?: string;
  phone?: string;
}

export interface ChangePasswordRequest {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface VerifyIdentityRequest {
  email: string;
  /** @minLength 1 */
  phone: string;
}

export interface VerifyIdentityResponse {
  verified: boolean;
  /** Short-lived JWT (15 min) required by /auth/owner/reset-password. Only present when verified is true. */
  resetToken?: string;
}

export interface ResetPasswordRequest {
  /**
   * Token obtained from /auth/owner/verify-identity
   * @minLength 1
   */
  resetToken: string;
  /** @minLength 6 */
  newPassword: string;
}

export type MembershipType =
  (typeof MembershipType)[keyof typeof MembershipType];

export const MembershipType = {
  weekly: "weekly",
  monthly: "monthly",
  quarterly: "quarterly",
  yearly: "yearly",
} as const;

export interface Member {
  id: string;
  gymId: string;
  name: string;
  phone: string;
  email: string;
  membershipType: MembershipType;
  startDate: string;
  expiryDate: string;
  lastPaymentDate: string | null;
  isPaid: boolean;
  paymentAmount: number;
  notes: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateMemberRequest {
  /** @minLength 1 */
  name: string;
  phone?: string;
  email?: string;
  membershipType: MembershipType;
  startDate: string;
  expiryDate: string;
  lastPaymentDate?: string | null;
  isPaid?: boolean;
  /** @minimum 0 */
  paymentAmount?: number;
  notes?: string;
}

export interface UpdateMemberRequest {
  /** @minLength 1 */
  name?: string;
  phone?: string;
  email?: string;
  membershipType?: MembershipType;
  startDate?: string;
  expiryDate?: string;
  lastPaymentDate?: string | null;
  isPaid?: boolean;
  /** @minimum 0 */
  paymentAmount?: number;
  notes?: string;
}

export interface RecordPaymentRequest {
  /** @exclusiveMinimum 0 */
  amount: number;
}

export interface MemberResponse {
  member: Member;
}

export interface MembersListResponse {
  members: Member[];
}

export interface StaffMember {
  id: string;
  gymId: string;
  name: string;
  username: string;
  createdAt: string;
}

export interface StaffLoginRequest {
  /** @minLength 1 */
  gymId: string;
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  pin: string;
}

export type StaffAuthResponseStaffRole =
  (typeof StaffAuthResponseStaffRole)[keyof typeof StaffAuthResponseStaffRole];

export const StaffAuthResponseStaffRole = {
  staff: "staff",
} as const;

export type StaffAuthResponseStaff = {
  id: string;
  gymId: string;
  name: string;
  role: StaffAuthResponseStaffRole;
};

export interface StaffAuthResponse {
  token: string;
  staff: StaffAuthResponseStaff;
}

export interface CreateStaffRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /**
   * Lowercase letters, numbers, and underscores only
   * @minLength 2
   * @maxLength 30
   * @pattern ^[a-z0-9_]+$
   */
  username: string;
  /**
   * 4–6 numeric digits
   * @pattern ^\d{4,6}$
   */
  pin: string;
}

export interface StaffListResponse {
  staff: StaffMember[];
}

export type CleanupExpiredMembers200 = {
  /** Number of member records removed */
  deleted: number;
};