# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gym Manager API specification
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Reusable schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
      required: [ok]

    # --- Health ---
...
    Owner:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gymId:
          type: string
          example: W8ASYS
        name:
          type: string
        gymName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
      required: [id, gymId, name, gymName, phone, email]
    RegisterOwnerRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        gymName:
          type: string
          minLength: 1
          maxLength: 100
        phone:
          type: string
          default: ""
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [name, gymName, email, password]
    LoginOwnerRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        owner:
          $ref: "#/components/schemas/Owner"
      required: [token, owner]
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        gymName:
          type: string
          minLength: 1
          maxLength: 100
        phone:
          type: string
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 6
      required: [currentPassword, newPassword]
    VerifyIdentityRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 1
      required: [email, phone]
    VerifyIdentityResponse:
      type: object
      properties:
        verified:
          type: boolean
        resetToken:
          type: string
          description: Short-lived JWT (15 min) required by /auth/owner/reset-password. Only present when verified is true.
      required: [verified]
    ResetPasswordRequest:
      type: object
      properties:
        resetToken:
          type: string
          minLength: 1
          description: Token obtained from /auth/owner/verify-identity
        newPassword:
          type: string
          minLength: 6
      required: [resetToken, newPassword]

    # --- Member ---
...
paths:
...
  /auth/owner/register:
    post:
      operationId: registerOwner
      tags: [auth]
      summary: Register a new gym owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterOwnerRequest"
      responses:
        "201":
          description: Owner registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/owner/login:
    post:
      operationId: loginOwner
      tags: [auth]
      summary: Log in as a gym owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginOwnerRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/owner/me:
    get:
      operationId: getOwnerProfile
      tags: [auth]
      summary: Get the current owner's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Owner profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Owner not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/owner/profile:
    patch:
      operationId: updateOwnerProfile
      tags: [auth]
      summary: Update the current owner's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/owner/change-password:
    post:
      operationId: changeOwnerPassword
      tags: [auth]
      summary: Change the current owner's password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Validation failed or current password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/owner/verify-identity:
    post:
      operationId: verifyOwnerIdentity
      tags: [auth]
      summary: Verify owner identity for password reset (step 1)
      description: >
        Checks email + phone combination. On success returns a short-lived
        `resetToken` (15 min) that must be passed to `/auth/owner/reset-password`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyIdentityRequest"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyIdentityResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/owner/reset-password:
    post:
      operationId: resetOwnerPassword
      tags: [auth]
      summary: Reset owner password (step 2)
      description: >
        Requires the `resetToken` issued by `/auth/owner/verify-identity`.
        The token expires after 15 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- Members ---
...