# lib/api-zod/src/generated/types/owner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Owner {
  id: string;
  gymId: string;
  name: string;
  gymName: string;
  phone: string;
  email: string;
}